refactor: 테스트 더블을 한개 디렉토리로 이동
