+++ src/test/java/com/github/thundermarket/thundermarket/ProductServiceAddTest.java
@@ -3,6 +3,8 @@ package com.github.thundermarket.thundermarket;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.thundermarket.thundermarket.TestDouble.ProductDetailFakeRepository;
+import com.github.thundermarket.thundermarket.TestDouble.ProductFakeRepository;
 import com.github.thundermarket.thundermarket.domain.Product;
 import com.github.thundermarket.thundermarket.domain.ProductDetail;
 import com.github.thundermarket.thundermarket.domain.ProductResponse;
