+++ src/test/java/com/github/thundermarket/thundermarket/TestDouble/ProductDetailRepositoryStub.java
@@ -1,12 +1,7 @@
-package com.github.thundermarket.thundermarket;
+package com.github.thundermarket.thundermarket.TestDouble;
 
-import com.github.thundermarket.thundermarket.domain.Product;
 import com.github.thundermarket.thundermarket.domain.ProductDetail;
 import com.github.thundermarket.thundermarket.repository.ProductDetailRepository;
-import com.github.thundermarket.thundermarket.repository.ProductRepository;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class ProductDetailRepositoryStub implements ProductDetailRepository {
